Do not reinstall Powerline font if it's already exist
Speed things up if the fonts is already installed.

+++ bootstrap.sh
@@ -13,7 +13,19 @@ function doIt() {
     -avh --no-perms . ~;
 }
 
+function is_powerline_fonts_exist() {
+  if ls ~/Library/Fonts/*Powerline.ttf 1> /dev/null 2>&1; then
+    return 0;
+  else
+    return 1;
+  fi
+}
+
 function install_powerline_fonts() {
+  if is_powerline_fonts_exist; then
+    echo "Powerline fonts already exist, skip this step."
+    return 0;
+  fi
   if [ -d "fonts" ]; then
     rm -rf fonts
   fi
